수정시 글 작성자와 안맞으면 동작 x

+++ src/main/java/com/example/pj1be20231109/controller/BoardController.java
@@ -49,7 +49,16 @@ public class BoardController {
     }
 
     @DeleteMapping("/remove/{id}")
-    public ResponseEntity remove(@PathVariable Integer id){
+    public ResponseEntity remove(@PathVariable Integer id, @SessionAttribute(value = "login", required = false) Member login){
+
+
+        if(login == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401
+        }
+
+        if(!service.hasAccess(id, login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
 
         if(service.remove(id)){
             return ResponseEntity.ok().build();
@@ -59,7 +68,16 @@ public class BoardController {
     }
 
     @PutMapping("/edit")
-    public ResponseEntity edit(@RequestBody Board board) {
+    public ResponseEntity edit(@RequestBody Board board, @SessionAttribute(value = "login",required = false) Member login) {
+
+        if(login == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if(!service.hasAccess(board.getId(), login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
 
         if (service.validate(board)) {
